Cache instance creation moved to separate module
