+++ core/scripts/server.js
@@ -3,8 +3,8 @@ const path = require('path')
 const express = require('express')
 const rootPath = require('app-root-path').path
 const resolve = file => path.resolve(rootPath, file)
+const cache = require('./utils/cache-instance')
 let config = require('config')
-const TagCache = require('redis-tag-cache').default
 const utils = require('./server/utils')
 const compile = require('lodash.template')
 
@@ -17,19 +17,6 @@ const isProd = process.env.NODE_ENV === 'production'
 process.noDeprecation = true
 
 const app = express()
-
-let cache
-if (config.server.useOutputCache) {
-  const cacheKey = JSON.parse(fs.readFileSync(path.join(__dirname, '..', 'build', 'cache-version.json')) || '')
-  const redisConfig = Object.assign(config.redis, { keyPrefix: cacheKey })
-
-  cache = new TagCache({
-    redis: redisConfig,
-    defaultTimeout: config.server.outputCacheDefaultTtl
-  })
-
-  console.log('Redis cache set', redisConfig)
-}
 
 const templatesCache = {}
 let renderer
